addTs
