+++ src/App.tsx
@@ -1,8 +1,6 @@
 import "./components/MainStyles/style.css"
 import MainRoutes from "./routes/MainRoutes";
 
-
-
 function App() {
   return (
     <div className="App">
