--- src/components/Cards/cards.module.css
@@ -1,45 +1,0 @@
-.card {
-    display: inline-block;
-    max-width: 300px;
-    max-height: 400px;
-    height: 100%;
-    width: 100%;
-    border: none;
-    box-shadow: 3px 4px 7px 3px rgba(0, 0, 0, 0.15);
-    border-radius: 10px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin-bottom: 20px;
-}
-
-.btn {
-    border: none;
-    background: transparent;
-}
-
-.card button a {
-    color: #fff;
-    text-decoration: none;
-}
-
-.card button:hover a {
-    color: rgb(16, 187, 135);;
-}
-
-.card h2{
-    margin-top: 10px;
-    font-size: 24px;
-}
-
-.card button {
-    display: inline-block;
-    margin: 0;
-    margin-bottom: 10px;
-    margin-top: 10px;
-    margin: 10px 25px;
-}
-
-.image {
-    max-width: 200px;
-}
