--- src/components/Green-Button/GreenButton.jsx
@@ -1,7 +1,0 @@
-import styles from "./greenButton.module.css"
-
-function GreenButton ({children}){
-   return <button type="submit" className={styles.directoryBtn}>{children}</button>
-}
-
-export default GreenButton
