--- src/components/Header/Header.jsx
@@ -1,46 +1,0 @@
-import { Link, NavLink } from "react-router-dom";
-import styles from "./header.module.css"; // Импортируем переменные стилей из модуля
-import { navs } from "./navLinks"
-function Header() {
-  return (
-    <header className={styles.header}>
-      <div className={styles.containerHeader}>
-        <div className={styles.logo}>
-          <Link to={'/'}><img src={require('../images/logotype.png')} alt="Logo" /></Link>
-        </div>
-        <nav className={styles.navigation}>
-          <ul>
-          {navs.map((item, index) => (
-            <li key={index}><NavLink  to={item.path}>
-              {item.title}
-            </NavLink></li>
-          ))}
-          </ul>
-        </nav>
-
-        <div className={styles.buttons}>
-          <button className={styles.buttonsIcons}>
-          <svg width="13" height="14" viewBox="0 0 13 14" fillRule="none" xmlns="http://www.w3.org/2000/svg">
-                    <path d="M6 0.7C4.04 0.7 2.46 2.29 2.46 4.25C2.46 6.2 4.04 7.79 6 7.79C7.95 7.79 9.54 6.2 9.54 4.25C9.54 2.29 7.95 0.7 6 0.7ZM3.71 4.25C3.71 2.98 4.73 1.95 6 1.95C7.26 1.95 8.29 2.98 8.29 4.25C8.29 5.51 7.26 6.54 6 6.54C4.73 6.54 3.71 5.51 3.71 4.25ZM2.66 9.04C0.94 9.04 -0.46 10.44 -0.46 12.16L-0.46 13.15C-0.46 13.78 -0.01 14.31 0.61 14.42C4.18 15 7.82 15 11.38 14.42C12 14.31 12.46 13.78 12.46 13.15L12.46 12.16C12.46 10.44 11.05 9.04 9.33 9.04L9.05 9.04C8.89 9.04 8.74 9.06 8.59 9.11L7.87 9.34C6.65 9.74 5.34 9.74 4.12 9.34L3.4 9.11C3.25 9.06 3.1 9.04 2.95 9.04L2.66 9.04ZM0.79 12.16C0.79 11.13 1.63 10.29 2.66 10.29L2.95 10.29C2.97 10.29 2.99 10.29 3.01 10.3L3.73 10.53C5.2 11.01 6.79 11.01 8.26 10.53L8.98 10.3C9 10.29 9.02 10.29 9.05 10.29L9.33 10.29C10.37 10.29 11.21 11.13 11.21 12.16L11.21 13.15C11.21 13.17 11.19 13.18 11.18 13.18C7.75 13.74 4.25 13.74 0.82 13.18C0.8 13.18 0.79 13.17 0.79 13.15L0.79 12.16Z" fillRule="#000000"/>
-                </svg>
-          </button>
-
-          <button className={styles.buttonsIcons}>
-          <svg width="16" height="15" viewBox="0 0 16 15" fillRule="none" xmlns="http://www.w3.org/2000/svg">
-                    <path d="M13.95 1.37L2.44 1.37L1.46 -0.39C1.35 -0.59 1.14 -0.71 0.91 -0.71L-0.34 -0.71C-0.68 -0.71 -0.96 -0.43 -0.96 -0.09C-0.96 0.26 -0.68 0.54 -0.34 0.54L0.55 0.54L1.52 2.28L3.76 7.23L3.96 7.67L1.71 10.07C1.55 10.24 1.5 10.49 1.58 10.71C1.66 10.93 1.86 11.09 2.09 11.12L4.14 11.35C6.7 11.64 9.29 11.64 11.85 11.35L13.9 11.12C14.25 11.08 14.49 10.77 14.45 10.42C14.41 10.08 14.1 9.83 13.76 9.87L11.71 10.11C9.24 10.39 6.75 10.39 4.28 10.11L3.47 10.02L5.12 8.26C5.14 8.24 5.15 8.22 5.16 8.2L5.79 8.28C6.67 8.4 7.56 8.42 8.44 8.35C10.5 8.19 12.41 7.21 13.75 5.63L14.23 5.06C14.25 5.04 14.26 5.02 14.27 5L15.17 3.63C15.81 2.66 15.11 1.37 13.95 1.37ZM5.21 6.95C5.07 6.93 4.95 6.84 4.89 6.72L4.89 6.71L3.04 2.62L13.95 2.62C14.12 2.62 14.22 2.8 14.13 2.94L13.25 4.28L12.79 4.82C11.67 6.15 10.07 6.97 8.34 7.11C7.55 7.17 6.75 7.15 5.95 7.04L5.21 6.95ZM3.41 12.41C2.72 12.41 2.16 12.97 2.16 13.66C2.16 14.35 2.72 14.91 3.41 14.91C4.1 14.91 4.66 14.35 4.66 13.66C4.66 12.97 4.1 12.41 3.41 12.41ZM12.58 12.41C11.89 12.41 11.33 12.97 11.33 13.66C11.33 14.35 11.89 14.91 12.58 14.91C13.27 14.91 13.83 14.35 13.83 13.66C13.83 12.97 13.27 12.41 12.58 12.41Z" fillRule="#000000"/>
-                </svg>
-          </button>
-
-          <button className={styles.constructor}>Конструктор
-          <svg width="13" height="13" viewBox="0 0 13 13" fillRule="none" xmlns="http://www.w3.org/2000/svg">
-                    <path d="M12.24 2.23C12.41 2.06 12.43 1.8 12.3 1.61C12.28 1.58 12.26 1.55 12.24 1.53L10.35 -0.36C10.26 -0.45 10.13 -0.5 10 -0.5C9.94 -0.5 9.89 -0.5 9.83 -0.48C9.76 -0.45 9.7 -0.41 9.64 -0.36L3.49 5.79C3.44 5.85 3.41 5.92 3.39 6L2.72 8.55C2.67 8.72 2.72 8.9 2.85 9.03C2.91 9.09 2.99 9.14 3.08 9.16C3.16 9.18 3.25 9.18 3.33 9.16L5.88 8.49C5.97 8.47 6.05 8.43 6.11 8.36L12.24 2.23ZM11.26 4.74C11.33 6.31 11.27 7.88 11.09 9.44C10.99 10.31 10.28 11.01 9.39 11.11C7.16 11.36 4.84 11.36 2.6 11.11C1.71 11.01 1.01 10.31 0.9 9.44C0.64 7.15 0.64 4.84 0.9 2.55C1.01 1.68 1.71 0.98 2.6 0.88C4.07 0.71 5.56 0.66 7.04 0.71C7.12 0.71 7.19 0.68 7.24 0.63L7.9 -0.03C7.98 -0.11 7.93 -0.25 7.82 -0.26C6.04 -0.36 4.25 -0.31 2.49 -0.12C1.15 0.03 0.07 1.09 -0.09 2.44C-0.37 4.8 -0.37 7.19 -0.09 9.55C0.07 10.9 1.15 11.96 2.49 12.11C4.8 12.36 7.19 12.36 9.5 12.11C10.84 11.96 11.92 10.9 12.08 9.55C12.3 7.7 12.35 5.83 12.23 3.97C12.22 3.86 12.08 3.81 12 3.89L11.34 4.54C11.29 4.59 11.26 4.67 11.26 4.74ZM10 0.7L11.17 1.88L5.5 7.56L3.9 7.98L4.32 6.38L10 0.7Z" fillRule="#000000"/>
-                </svg>
-          </button>
-        </div>
-      </div>
-      <hr />
-    </header>
-  );
-}
-
-export default Header;
