--- src/components/addInput/AddInput.jsx
@@ -1,13 +1,0 @@
-import React from 'react'
-import styles from "./addInput.module.css"
-const AddInput = ({type, placeholder,name,onChange}) => {
-  return (
-    <div className={styles.inputBlocks}>
-        <label className={styles.label}><p>{name}</p>
-        <input type={type} onChange={onChange} placeholder={placeholder} className={styles.input}/>
-        </label>
-    </div>
-  )
-}
-
-export default AddInput
