--- src/components/addInput/addInput.module.css
@@ -1,34 +1,0 @@
-.input {
-    font-size: 14px;
-    margin-bottom: 10px;
-    min-width: 250px;
-    box-shadow: 0px 4px 7px 0px rgba(0, 0, 0, 0.15);
-    border-radius: 10px;
-    border: none;
-    padding: 12px 15px 12px 15px;
-    display: inline-block;
-}
-
-.input:focus-visible {
-    outline: 0;
-}
-
-.label {
-    max-width: 500px;
-    width: 100%;
-    display: flex;
-    justify-content: space-between;
-}
-
-.inputBlocks {
-    max-width: 430px;
-    width: 100%;
-}
-
-.label p{
-    padding-bottom: 10px;
-    padding-top: 10px;
-    display: inline-block;
-    font-weight: 500;
-
-}
