+++ src/index.tsx
@@ -1,12 +1,12 @@
-import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
 import { BrowserRouter } from 'react-router-dom';
+import App from './App';
+import React from 'react';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
     <BrowserRouter>
-    <App />
+    <App/>
     </BrowserRouter>
 );
