+++ src/pages/MainPage/AboutBlock/AboutBlock.tsx
