+++ src/pages/MainPage/MiniInfo/MiniInfo.tsx
